TXNCXX-11 Implement API
Motivation
==========
All we want are some initial tests, to consider this done.  These are
very simple client tests only.  More to come, but this is a start.

Modification
============
A little CMake work to pull in gtest, and a set of very simple tests
of cluster, bucket and collection.  Right now we require a running
server but probably more appropriate to use the mock, later.

Results
=======
A single executable (client_tests) which one can run from the build
directory.  Ideally I'd have a make target for this, but there's a bit
of voodoo necessary to get the gtest_add_test stuff working.  Will definitely
do that soon.

Change-Id: Ia9d71d13cd2a8308cd031ae3cc6dac89c82c5d69
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/132586
Reviewed-by: Sergey Avseyev
Tested-by: David Kelly <davidmichaelkelly@gmail.com>