Performer bug fixes (round 1)
Motivation
==========
Running the StandardTest suite on the fit performer, there were
many failures.  This represents a first set of changes, that get
16 of the 35 tests to pass.  The remainder require implementation
of rollback, use of links, and forcing errors, which we can do
next.

Modification
============
Mostly getting the transaction to return a TransactionResult which
can be used by the fit performer.  Also, added get_optional, and a
number of small fixes for things which didn't match the specs or
expectations, given that they have changed since this was initially
implemented.  Also - some things will be looked at again - in particular
the transaction_links where, to avoid some copy issues, I eliminated
the const-ness of the members of the class.

Results
=======
No new tests, but the existing client tests all pass.  Worth pondering
a set of transaction unit tests, but not right now.

Change-Id: I3f7e525cc802aae283bc23b1f3af3773a57d424c
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/133053
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>