Another cmake issue
Though happy to build on mac, on linux we see an issue with gtest,
as libcouchbase has a copy, yet this project just downloads it and
builds it.  So now, we just use the libcouchbase copy.  Also removed
references to gmock, since that isn't in libcouchbase. Ideally, we'd
update LCB to download it (so we can update the library easily), and
add gmock, etc...

But for now, I just reworked a couple places in the tests where I used
EXPECT_THAT (a gmock thing), and we are good.

Change-Id: I6c1231fb35930772f287e780d2686541b032c19c
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/133083
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>