More cmake changes, plus some minor code things
Since we need to build on jenkins, I changed the cmake a bit.  Now
we download and build spdlog ourselves, since cbdep doesn't have it.
However, boost seems to be available, so for now lets assume that is
installed.  For jenkins we can cbdep it, so not a problem probably.

Boost needs to be beyond 1.57 or so - as has_value() was added to
boost::optional around then. If this becomes problematic, we can add
it like we did spdlog.

Also, linux was a bit more restrictive, and forced a couple small
changes to the code, which works fine on the mac as well.

Change-Id: If3e7e82f5dca0bb5560116871bf7384fab84c647
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/133163
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>