TXNCXX-16 Implement Rollback
Motivation
===========
Need to roll things back.

Modification
============
Implemented rollback logic, following java/specs as closely as I
could.  While there, put in some of the retry logic, and the errors
are now using and ErrorMapper per the spec (that is TXNCXX-24).

Some caveats:  we don't do tombstones yet for inserts, so a couple
StandardTests don't pass in the fit performer.  Also, the existing
commit logic doesn't call the hooks it should.  Will rework that in
a separate ticket, as part of some other work.  There is more retry
logic to put in, and for that matter more error handling as well.

Results
=======
No new client tests, but they all pass, and 25 of 35 StandardTests
on the fit performer side of things now pass as well.

Change-Id: Id613e28b9b9259cf2ea2d638c2ee8a523591a8d3
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/133922
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>