TXNCXX-23 Retry loop plus final exceptions
Motivation
==========
Implement final exceptions and retry loop per specs

Modifications
=============
Main thing here involved tweaking the exceptions and exception handling so
we retry or rollback when we should.  Also, added some hook calls and the
testing exceptions.  Gets StandardTest.java to pass 27 of 35 tests.  Remainder
need some other hooks and related code to be added.

Results
=======
We are now retrying when a retryable exception is raised, and several tests which
rely on that are passing.

Change-Id: If53168c9d606bcb6705d976584eeee5907a4580e
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/134223
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>