TXNCXX-29 Intermittent client test failures
Motivation
==========
Reliable tests

Modification
============
First, we now log with spdlog in tests.  A bit better readability
in Jenkins.  Also - locally using cbdyncluster, I could reproduce
the intermittent issues (not with a local instance, or local cluster).

So we now cache the bucket and collection for all the collection
tests.  That helped a bit, then, in a fit of lazyness, a 10 second
sleep whenever we get the bucket and collection in the setup seems
to avoid the issue.  When we find the root cause (further down the
stack), I will take great pleasure in removing the sleep.

Results
=======
Passing tests on Jenkins.  Probably that 10 seconds could be needing
a tweak at some point, but we will see...

Change-Id: I6f01c43139725d856bf1e0df1b9d3bb247e2ca3d
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/134283
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>