TXNCXX-35 Version strings in client and transactions
Motivation
==========
We need the version burned in, so we can diagnose issues now that
a release is coming up.  For now, no packaging, lets just be sure
that the version is echo'd in the logs.

Modification
============
CMake has the version, writes support.hxx with it in there.  We have
a VERSION_STR macro we can call as well.  CMake pops the git sha in
as well.

Results
=======
You can see it echoed in the tests when you create a new cluster.  Same
if you create a new transaction object.

Change-Id: I1f0fe0877c5cd5d297079216145b242b4f7b2229
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/134691
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>