TXNCXX-35 TXNCXX-36: Versioning updated, operator<< fixes
Motivation
==========
We really should version the shared lib, and those operator<< are super
useful to anyone using the library (as I discovered when I started creating
a standalone example app).

Modification
=============
Updated CMakeLists.txt to add the versioning to the .so.  That means it is built
with the version in the filename in the standard way.  Updated the process a bit,
doing more work in cmake, less in the header.

Fixed the transaction_document's operator<< which didn't work with spdlog, and
eliminated the to_string on client::result in favor of same style operator<<.
This now works for any ostream (like cout, etc, spdlog, etc...)

We can add more of these, for other objects, as needed.

Result
======
No test changes, but more ready to go out the door.

Change-Id: If060133f9e47cc71e46566136e0ca493a85e8c36
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/134856
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>