TXNCXX-39 Stage content properly
Motivation
==========
The content during a remove or replace must be staged where expected
to handle cases where a txn is left in a bad state (say, a crash).
Also - there are issues with the metadata not being propogated into
the ATR in some cases, leading so similar issues.

Modification
============
The metadata just needed to be copied in the transaction_document's
copy constructor.  The staging of content during replace was not done,
simple to add.  Also we _don't_ expect staged content for a remove but
there was a sentenel in place, which was removed.

Result
======
More StandardTest passes (only 5 fails now)

Change-Id: Idea2e18a2bc9bfaad56396c7030973f289b29721
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/135375
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>