TXNCXX-42 Really terminate when expired
Motivation
==========
Fix a small bug where we were not rolling back, and raising the correct
exception, when expired after multiple attempts.

Modification
============
Fixed the attempt_expired exception, adjusted the transactions.run call
to not check for expiry (that is done throughout the code elsewhere), and
extended the max_retries to not hit that before timing out.

Result
======
The ErrorTests no longer fail for this reason.  4 still fail, but that is because
they need some fit performer work (the MUTATE_DOC hook action, etc...).

Change-Id: I03aaa299d75906f9b09dd861dc6e05be9f707d95
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/135803
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>