TXNCXX-45 Error Handling rework
Motivation
==========
Since the client does not raise exceptations, but the testing in
fit performer does (in hooks), a simple way to handle this with little
extra code would be helpful.

Modification
===========
Wrapping the calls to collection->xxx, and having it throw an error that
classifies the result code allows for simple handling of test exceptions.
Also - tweaked some error handling to match java.  More needed, but it is
a start.

Result
======
The ErrorHandlingTests now pass.

Change-Id: I2979a7ead0b5e9f0d62df86e1e4fe341b7393b08
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/136411
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>