TXNCXX-54: Test Crashes
Motivation
==========
Some state/* tests segfaulted.  Also - worth sweeping through and fixing
some of them, and adding to jenkins.

Modifications
=============
Some rework of retry handling (fixing stack corruption due to infinite
loop of retries).  Also implemented the check_atr_entry_for_blocking_document
call, so we better handle write-write conflicts.  More to do there, but
this is a big part.  Some error handling fixes (thx to state tests finding
a number of issues).  Again, more to do, but timeboxed this for now.

Result
======
About 100/300 state tests pass -- more to do there but it is a start.  Not
cleaning up accounts for some of them, a number of them stem from a couple
other bugs but time to move on.

Change-Id: Ib35ff68ba16a4a3f9cf57fe8e156d8f61b84d55b
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/136974
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>