TXNCXX-20 TXNCXX-62 Cache buckets, fix mutate_in_spec::remove
Motivation
==========
Cleanup requires us to referernce the atr collection in the cleanup request,
and the associated bucket needs to be created and cached.  The cluster for the
cleanup thread is already available, makes sense to do this caching now.

Also - noticed issue with doc flag during mutate_in when removing - simple fix
(thanks to Sergy)

Modification
============
Cluster keeps buckets it opened in a vector.  When you ask for one, it gives
you one in that vector if the name matches.  Yea... linear search but we dont
expect 1000s of open buckets.

The mutate_in_spec just needed to only set the store semantics on upsert and
insert - not on remove.

Results
=======
Added subdoc tests, all pass.  Added a test for caching as well.  Need more
subdoc testing, but it is a reasonable start.  Will add jira ticket for that.

Change-Id: I7d6e29c8c8fe37a951d9c18389c115f3deafd4c0
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/137418
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>