TXNCXX-72 cache collections in bucket
Motivation
==========
Efficiency, why alloc/dealloc all those objects?  Also - if we
don't cache them, the collection reference in the staged docs can
be already deleted.  Of courde that could be a shared_ptr as well.
But in the end this seems sensible, regardless of anything else.

Modification
============
Just a linear search to find it in a vector for now, but we can
consider a sorted approach if needed.  Note it is scope and name
that would need to be sorted, so maybe not worth the effort until
we see a reason.

Results
=======
Added test for it, slight modification of the cluster test for
cached buckets as well.

Change-Id: If687b589fadd17f6711a995f6f8e933575ac0a5a
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/137860
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>