TXNCXX-75 Matching error_class
Motivation
==========
The original issue in TXNCXX-75 pointed out that my error_class was
not ordered the same as the fit performer.  That complicates the
comparisons, needlessly.  We also look at attempt_state - so made
sure it starts at 0.

Modification
============
Reorder to match the protobuf ErrorClass in fit performer - no impact
here as we always use it as an enum.

Result
======
Fit performer comparisons now work (after that is checked in).

Change-Id: I9527b5760c68272c0545d826a13edfcef8787398
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/138201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>