TXNCXX-60 Lost attempts cleanup
Motivation
==========
Crashes and bugs could leave attempts in atrs that don't get cleaned
as they never make it into the cleanup queue.  So, we need another
mechanism to just sweep through all atrs and clean them up.

Modification
============
We now spin up a thread which periodically spawns a thread per bucket,
and finds any atrs in those buckets and cleans each attempt, if that
attempt has expired.  Along the way, I implemented collection.exists,
as it is more efficient (I think) than a get for what may be a sparse
set of keys.  That is debatable, of course...

Results
=======
New tests for exists.

Change-Id: I67095c2e25af936617045df3ab2a241080e45d10
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/138441
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>