TXNCXX-60 Lost cleanup loop bugs
Motivation
==========
Fit performer uncovered some immediate issues in the previous lost
cleanup loop work.

Modification
============
We now remember to pass the cleanup state (via transactions_cleanup_attempt)
by passing it in, when we are given one.  And the main thing, I
got rid of the atr_cleanup_entry constructor that the fit performer
(and nobody else) uses.

Result
======
The tests that I know used to pass that depended on the cleanup now
will pass again.

Change-Id: I4627edaad22018f6ddbc84b088384fb5d643c3e1
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/139008
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>