TXNCXX-82 fails in Get, Replace, Remove, Insert state tests
Motivation
==========
Error handling issues in get, replace, remove, insert need some
fixing.

Modification
============
One issue was retrying - we could blow the stack depending on how
many times we did it.  Eliminated unbounded recursive retry in
favor of a retry wrapper around an object.  This fixes TXNCXX-81.
Currently a fFixed retry delay each time around, can make more
sophisticated later, as needed.

A second issue is a workaround for CCBC-1323.  For now, we workaround
by assuming the LCB_ERR_DOCUMENT_EXISTS in a mutate_in call is always
an LCB_ERR_CAS_MISMATCH.  When CCBC-1323 is fixed, we can remove this.

The rest are a series of small fixes in error handling mostly, so we
can pass all the tests mentioned above.

Results
=======
Get, Replace, Remove, and InsertTest in states are all passing.

Change-Id: If3404e3037339618d145b0dd2fae4856b997e39b
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/139282
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>