TXNCXX-85 Cleanup tests
Motivation
==========
Getting cleanup tests in fit performer to pass.  However, while at
it, we needed to get to the bottom of the issue we saw with transactions
committing replaced documents which were out of date.

Modification
============
When unstaging documents, the xattrs had a null for the txn key, which
oddly enough made the subsequent mutate_in calls (which only manipulated
the txn xattr) succeed, but return 0 for the cas.  Thus when we committed
it, it _always_ succeeded, even when the document had changed in the meantime.
Simple fix there.

Also - changed logging to accept the attempt context (making debugging easier),
which we'd planned to do later but needed for the above.

Started work on TXNCXX-85 as well, but not finished.  Will continue to fix
cleanup issues in a coming commit.

Results
=======
Concurrency issue fixed, some of cleanup tests now passing.

Change-Id: I52a4ea678c1effe1cbbfdaee3f93f3d2019f607c
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/140449
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>