TXNCXX-87 Subdoc errors not always surfaced
Motivation
==========
We rely on LCB_ERR_SUBDOC_PATH_NOT_FOUND and others to be raised
in the transactions code.

Modification
============
Client now returns a vector of the value and status for each of the
specs in the mutate_in/lookup_in call.

Transactions code has wrapper that raises exceptions, and now it also
raises the _first_ of them in the values, if the request itself was
successful.  Mirroring java, more or less.

Added one handler for the PATH_NOT_FOUND, and changed some inserts
to upserts in another place, both more closely matching java and needed
to pass some fit perfomer test.

Also, since the wrap_collection_call needed to change, and was repeated
in a couple places, we now have it in one spot.  That fixes TXNCXX-73.

Results
=======
New test for client pass, and some failing fit perfomer tests pass now.

Change-Id: I1d0285a6bcbff817fc862fa4fc7873e3348a79cf
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/140662
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>