TXNCXX-85 Get cleanup tests passing
Motivation
==========
Finish implementation of cleanup

Modification
============
Client needs to only raise errors like PATH_NOT_FOUND when doing a
mutate, so needed this logic in place for some of the cleanup test
scenarios.  The client record needed to return information for testing
so added a client_record_details struct, and logic to fill it out
appropriately.  Cleanup hooks for client record needed to be added,
and error handling during the client record read/update needed work
as well.

Result
=====
All the client record cleanup tests pass, with the exception of the
override test.

Change-Id: I747a7071220b474f6fa94fb37b80d04d1f407614
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/141248
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>