TXNCXX-88 Thread-safe client and txns
Motivation
==========
This is probably the way most users would like to use transactions.  They
can create a single instance, use it in multiple threads.  And of course
the client should work the same.  You can use cluster, bucket, and collection
objects between threads.  This has benefits - the cleanup threads
can now stay with the transaction instance (no need to make them static), as
you don't need to make one transaction (and cluster, etc...) per thread.

Plus, users can manage connections by managing the number of LCB instances
in the pools.  You can manage the number of instances in the cluster and all
bucket pools by passing those in the cluster_options.

Modification
============
An internal pool holds on to instances in the cluster and bucket classes.  For
each operation, we grab an available one (or create a new one if all are busy),
up to the max instances.  When we are out of available instances, we wait for
one to become free, and use it.  You can specify the sizes in a new options
object - cluster_options -- when you create the cluster.  We probably will make
a set of authenticator options as well, as we evolve the client.

Results
=======
Added simple threaded tests for the client, and transactions, and pool.  Will expand
to have a small set of transaction tests that cover transactions a bit better, but
mostly we rely in fit performer for this.  Nice to have some level of testing outside
fit, however.

Change-Id: Ia15b011e14adb57d93e9efa2ada402bfc4de92d5
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/141801
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>