TXNCXX-95 Issue with deleting buckets
Motivation
==========
We are not releasing the LCB instances if the bucket destructors are
not called.  That's not good.

Modification
============
Using a std::weak_ptr reference to the bucket in collection does the
trick.  However, added a mechanism to see when instances are created,
destroyed, removed, added, etc... And some tests for that as well.

Results
=======
The Pool is tested much more thoroughly as a result.  Also, have a test
for this in the SimpleClientClusterTests, which insures instances are
created, destroyed, added and removed as expected from the cluster
and bucket pools.  Will add more tests to transactions, incuding
pool tests, soon.

Change-Id: I1e0e85a95b998eea791f834fe1f9705c4dcacdb9
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/142314
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>