TXNCXX-97 Set cause in transaction_operation_failed exception
Motivation
==========
Sometimes, (FAIL_DOC_ALREADY_EXISTS, FAIL_DOC_NOT_FOUND), we can go
ahead and set the cause accordingly, by default.  That helps us get
some fit performer tests to pass.

Modification
============
Add a simple function to translate error_class to external_exception
when it is clear what to do.  There are probably a number of other
places in code where we need to set this explicitly, which the fit
performer can find.  But this fixes some tests, and is a reasonable
default.

Result
======
MultipleTransactionsTest suite passes (with the help of TXNCXX-94
changes in the performer).

Change-Id: I0fa406f5409f653c90bef92831af852206d58026
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/142466
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>