TXNCXX-99 Cache errors from operations
Motivation
==========
Allows transactions to function properly even if users catch our
transaction_operation_failed exceptions.

Modification
============
We use a wrapper to call the operation, cache a transaction_operation_failed
exception if thrown before re-throwing it.  Also, if the cache isn't empty,
the wrapper throws appropriately.  The commit also throws if there is a
cached exception.

Result
======
The fit performer tests that rely on this now pass.  Note that there
is a small change in performer as well.

Change-Id: I87c851704827d5a02daa38bff8e3995d4463ed90
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/142581
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>