TXNCXX-110 Crash in ConcurrencyTest
Motivation
==========
Lets be sure we are thread-safe.

Modification
============
When we generate uuids, we use boost.  However the generators are
not completely thread-safe.  Simple enough to fix.

Results
=======
I have a benchmark that just spawns lots of threads and conflicts
trying to update docs in a txn.  I _did_ see a crash - and now it
seems to be fixed.

Change-Id: I8f55bc6231f00b81b5dd180cd7aa5a8ba80ebd74
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/142667
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>