TXNCXX-117  Handling timeouts
Motivation
==========
Insure we expose default timeout in transaction_config, and also
that the per-operation timeouts in the options are used when they
are present.

Modification
============
In the client, insured the per-operation timeouts specified in the
options are used, and exposed a default kv_timeout in the cluster_options.
This sets the default, plus insists that the durability floor is at least
as long.  These defaults are inherited by all instances in the cluster.

In transactions, exposed a kv_timeout which we use for all operations.
Introduced a wrap_option function that handles that (and the durability,
when appropriate) throughout the txn code.

While touching some headers, added a bit of api docs per TXNCXX-114, but
more to do on that in its own commit.  Additionally, hardened the tests
to situations where the bucket is not found.

Results
=======
New tests for client default kv_timeout and options, all pass.

Change-Id: Ief90922df69f4219d7404c3ef134368b85c15c0f
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/145153
Tested-by: Build Bot <build@couchbase.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>