TXNCXX-118 Retry until timeout when cluster/scope missing
Motivation
==========
This should be done in the lcb retries (CCBC-1373).  But for now, lets
workaround it by adding to the NO_BUCKET workaround (CCBC-1300).

Modification
============
Now we need to retry on LCB_ERR_COLLECTION_NOT_FOUND or LCB_ERR_SCOPE_NOT_FOUND,
in addition to the existing retries.  Also have to keep track of the deadline
and modify timeout.  That means, properly propogate the default kv timeout down
from the cluster, using it when no timeout is specified in options.

Also - turns out we were not very careful about how we parse the collection and
scope in bucket.collection(...), so added to that, with tests.  Oddly enough, the
collection cache in bucket was not really threadsafe, so added that.  May ponder
whether or not we need to cache them.

Tests now look for SUPPORTS_COLLECTIONS environment variable, to run some 7.0
specific tests.

Results
=======
Added tests pass, and MultipleCollectionTests in fit performer passes as well.

Change-Id: I5e9840f9342d9d693a36db4b4aa9fa4c32f39078
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/145999
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>