TXNCXX-120 Pin to lcb 3.0.7
Motivation
==========
Get up-to-date in anticipation of coming telemetry.

Modification
============
In 3.0.6, you could set the collection and scope to _default, without
issue in a < 7.0 cluster.  In fact, you could set them to anything you
want.  In lcb 3.0.7, setting them to anything results in a UNSUPPORTED_FEATURE
error in < 7.0, as it should.  So, don't set it to _default._default at
all, and you avoid all this.  Of course, you get the expected error if
trying to set to non-default on a < 7.0 cluster.

Eliminated use of boost's value_or (later boosts got rid of it, seems wise
to remove now).

Results
=======
Tests pass on < 7.0 and 7.0 clusters.

Change-Id: If40049b65636c08a230ab25ee333304e3073f9b9
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/145963
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>