TXNCXX-119 Eliminate ROLLED_BACK and COMPLETED states in atrs
Motivation
==========
Per CBD-3638, lets do less writing to the ATR, and cleanup successful
commits and rollbacks inline.

Modification
============
We no longer write those states, and instead we just remove the atr
entry.  And of course, no need to add them to the cleanup queue.

While doing this, I noticed I'd not converted some retries to the default
exponential backoff with jitter, so added that.  Also, when we remove
our client record, if we are the last record, the entire record xattrs
was being removed.  This makes tests fail subsequently, so now a simple
mutate_in_spec::upsert(path, nullptr).xattr() is in the spec, and that
seems to avoid it.

Results
=======
Every test passes on 6.6, with the exception of one, and I believe that
test is broken and will patch that on the fit performer side.

Change-Id: I1de061df22081b7583a0d0c4cfc903115aaddce2
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/146867
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>