Release Candidate rc1
Motivation
==========
Update build/cmake to 1.0.0.rc.1

Modification
============
Updated version in the usual places. Also, two of the recent utils
tests could fail occasionally, so tightented them up and commented
on them in the test for clarity.

Result
======
Builds now should be rc1, both externally (in the packaging), and
internally (in what it reports).

Change-Id: I558658013a830940a9968e78f5cdbb16c03dd5b2
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/147350
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>