TXNCXX-126 Min transaction retry time should be 1ms
Motivation
==========
Match the java implementation, which is the same as the RFC.

Modification
============
Switched to using the utils.hxx exponential backoff.  I'd been using them
everywhere else, just not here until they had some cycles of tests.  Now
they have unit tests plus plenty of exercise, so time to switch.  And, while
at it the min_retry_delay_ is now 1 ms, and the max_attempts_ was bumped
to 1000, as it was 100 (and is in java), as 100*128ms (the max retry delay)
is not even quite the default 15 second expiration time.

Also tweaked the logic surrounding arbitrary exceptions arising in the logic
outside our operations, and added tests.

Results
=======
Fit performer tests and unit tests all passing.

Change-Id: I267d76f5b2084487337d64c1628691c180352359
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/147482
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>