TXNCXX-132 Durable remove fix
Motivation
==========
Update to use durability when removing staged docs, now that libcouchbase
fix for that is in.

Modification
============
Bumped the libcouchbase sha to the CCBC-1383 fix sha.
Added a simple test, and updated the relevant staging mutation call.  While
at it, would like the LCB version echoed when making initial connection, so
updated that log in cluster.

Results
=======
Tests pass, including new one.

Change-Id: I85da7a276171466d85b7f71dade66c4d16c94438
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/150268
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>