TXNCXX-140 Minor API alignment with spec
Motivation
==========
Matching the spec for naming, etc...  Also, refresh the license file.

Modification
============
The transaction_document becomes transaction_get_result.  Also the
transaction_result has some unnecessary members.  While we are at
it, had info-level logging when we create a bucket, with the max
instances logged as well.  Similarly, when we create a transaction,
I log the sha in addition to the version.  And, replaced the license
file with the current APLv2 text.

Results
=======
Tests all pass, logging looks good, etc...

Change-Id: I2efaa3fa1d82a406fc4a20f0da387a50dcec4868
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/151326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>