TXNCXX-154 More async txn work
Motivation
==========
Found issue in fit tests with cleanup results

Modification
============
The transactions_cleanup_attempt struct mistakenly used a const ref to
the atr_id, which is on the stack only when there are no errors in the
cleanup.  Simple fix

Also, forgot to fixup the cleanup hooks to throw properly (for now), in
the cleanup (which is synchronous for now).

Results
=======
cleanup fit tests should pass (at least, the vast majority of them)

Change-Id: I8751b9a3907341a765aa0bd24462fcd775bf7743
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/167452
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>