TXNCXX-167 Update internal interfaces
Motivation
==========
Integration with some of the wrappers seems to necessitate an interface
much like gocbcore-transactions.   Lets provide that internal interface
for them.

Modification
============
The transaction_context now delegates the async attempt context
operations to the current attempt.  The async version of
transactions.run now is really async, and accepts a lambda for the
logic, then a second as a callback when the transaction is complete.

While at it, updated to latest client, corrected a couple issues that
came up when doing so (having to do with fmt lib),

Also, we have async versions of commit and rollback, though the need
for them at all is up for debate.

Results
=======
Updated unit tests all pass.

Change-Id: I419db2740c35e9600ab2039c81b9f41227ed40b5
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/167928
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>