TXNCXX-179 Scoped Query Support
Motivation
==========
Add support for scoped queries

Modification
============
Updated transaction_query_options accordingly.  While at it, re-worked
how we wrap the options to produce a request which has the correct
defaults and correct timeout fields, given the current config.   Need
this as part of the move to per-txn configs.

Also, added a test that fails and rollback, strictly because of the
logic, on the query side.  Ensured that we were handling the returned
errors from query correctly (though, not complete coverage, we let fit
do that).

Results
=======
All tests pass

Change-Id: I501f5ccbf9c77080cf0862ef2c5fd4b98e064b67
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/168633
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>