TXNCXX-163 Per-Transaction Configs
Motivation
==========
Adding the per-txn configs, as it is part of the spec now.

Modification
============
Added a class for the per-txn config, a subset of all the config
options.  It can be specified in the transactions.run call, or in the
constructor for the transaction_context.  Added a simple test to assure
us it actually is applied in the transaction_context constructor, which
is also used by transactions.run.

Also insured we converted expiry in config to ns.  Seems it was
implicitly doing so after a cast to the milliseconds.

Updated async tests to deal with the occasional transient error, as we
seem to have run into to some on jenkins.

Results
=======
Tests pass, with the one new one.

Change-Id: Idf23113e94575c1994542f2b9fafcbde634ea496
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/168764
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>