TXNCXX-172 Core loop logic now in transaction_context
Motivation
==========
Wrappers that use transaction_context can leverage this, eliminating
need to repeat the code that we use internally.

Modification
============
The transaction_context now has a finalize method, which can be called
after the transaction logic lambda, which waits for the operations to
complete, then will rollback, commit, retry as needed.  Also, there is a
poorly named handle_error, which finalize uses, and is also used around
the transaction logic, in case the lambda raises.  However, that
_should_ only be necessary for synchronous calls.

The transaction_context tests include a couple tests that use this, and
the transaction.run implementation uses it as well, so it will be well
tested and the same logic that the library uses internally.

Results
=======
All tests pass.

Change-Id: I8e2b1984ef246b6f89d3c6edc425b24b6b5467cc
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/168945
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>