TXNCXX-191: Add support for BF-CBD-3705
This bugfix handles the ATR entry being deleted, which is
most likely caused by cleanup having happened.  To check
for sure, we reread the document.  If it's still in the same
transaction, we know it was a Pending transaction that cleanup
has now handled.

Part of this bugfix is also to handle an ATR being deleted in the
same way, so that state does not block things.

Change-Id: Ifc32a71b236e62f37ab57a473cee68e7aaad378b
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169337
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>